Added generalized binomial Fixed dependencies

+++ src/quantum.rs
@@ -123,7 +123,7 @@ pub fn radial_wavefunction(n: usize, l: usize, r: f64) -> f64 {
     norm *= basic::factorial(n - l - 1) as f64 / (2 * n * basic::factorial(n + l)) as f64;
 
     // Computing the term associated to the Laguerre polynomial
-    let poly: f64 = polynomial::Poly::laguerre(n - l - 1, 2 * l as i32 + 1).compute(2.0 * factor);
+    let poly: f64 = polynomial::Poly::laguerre(n - l - 1, 2.0 * l as f64 + 1.0).compute(2.0 * factor);
 
     // Finishing computation and returning
     (2.0 * factor).powi(l as i32) * norm.sqrt() * poly * (-factor).exp()
@@ -145,7 +145,7 @@ pub fn radial_wavefunction(n: usize, l: usize, r: f64) -> f64 {
 pub fn radial_vec(n: usize, l: usize, r: &[f64]) -> Vec<f64> {
 
     // Preparing the poly
-    let poly = polynomial::Poly::laguerre(n - l - 1, 2 * l as i32 + 1);
+    let poly = polynomial::Poly::laguerre(n - l - 1, 2.0 * l as f64 + 1.0);
 
     // Computing the norm of the function
     let mut norm: f64 = (2.0 / (n as f64 * cst::A_0)).powi(3);
